Change callback example
I don't seem to have the synth function in the latest master of cl-collider. 
Made some quick changes to the example code to get it to run. 
Is this function included with cl-collider, or should I be acquiring it from somewhere else? 
Great tutorial btw, always good to see this library becoming more accessible/documented.